fix(clients/java): handle inactive role in Java client
