+++ clients/java/src/test/java/io/zeebe/client/TopologyRequestTest.java
@@ -20,6 +20,7 @@ import static io.zeebe.client.util.RecordingGatewayService.partition;
 import static io.zeebe.gateway.protocol.GatewayOuterClass.Partition.PartitionBrokerHealth.HEALTHY;
 import static io.zeebe.gateway.protocol.GatewayOuterClass.Partition.PartitionBrokerHealth.UNHEALTHY;
 import static io.zeebe.gateway.protocol.GatewayOuterClass.Partition.PartitionBrokerRole.FOLLOWER;
+import static io.zeebe.gateway.protocol.GatewayOuterClass.Partition.PartitionBrokerRole.INACTIVE;
 import static io.zeebe.gateway.protocol.GatewayOuterClass.Partition.PartitionBrokerRole.LEADER;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
@@ -67,7 +68,7 @@ public final class TopologyRequestTest extends ClientTest {
             432,
             "3.22.3-SNAPSHOT",
             partition(0, FOLLOWER, UNHEALTHY),
-            partition(1, FOLLOWER, UNHEALTHY)));
+            partition(1, INACTIVE, UNHEALTHY)));
 
     // when
     final Topology topology = client.newTopologyRequest().send().join();
@@ -115,7 +116,7 @@ public final class TopologyRequestTest extends ClientTest {
         .extracting(PartitionInfo::getPartitionId, PartitionInfo::getRole, PartitionInfo::getHealth)
         .containsOnly(
             tuple(0, PartitionBrokerRole.FOLLOWER, PartitionBrokerHealth.UNHEALTHY),
-            tuple(1, PartitionBrokerRole.FOLLOWER, PartitionBrokerHealth.UNHEALTHY));
+            tuple(1, PartitionBrokerRole.INACTIVE, PartitionBrokerHealth.UNHEALTHY));
   }
 
   @Test
